[FIX] changed prototypes for map_get_type

+++ src/include/map.h
@@ -19,7 +19,7 @@ struct map
     struct vector2 start;
 };
 struct map *load_map(const char *filename);
-enum block_type map_get_type(struct map *map, size_t i, size_t j);
+enum block_type map_get_type(struct map *map, double i, double j);
 void map_print(struct map *map);
 
 #endif /*MAP_H_*/
